requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.0
Django==1.11
django-celery-results==1.0.1
djangorestframework==3.8.2
kombu==4.2.1
pytz==2018.5
selenium==3.13.0
vine==1.1.4
XlsxWriter==1.0.5
